FEATURE - implement KeyboardController and DirectionKeyEvent
The Keyboard Controller will monitor user keyboard input, and send events to the Internal Controller Event Handler when notable keypresses are made.
Currently, the only event planned is a DirectionKeyEvent, which will only occur if the arrow keys are being held down in a manner that points in a certain direction.