FEATURE - implementation of Mouse Controller which supports SelectionEvent/MouseOrderEvent
The Mouse Controller handles mouse input by the user. This version currently supports single unit selection via left click, and mouse command via right click. Area click-and-drag selection is not available, but a coroutine has been defined to support this in the future.

The Mouse Controller will sent a SelectionEvent to the event handler after a left mouse click, and a MouseOrderEvent to the event handler after a right mouse click.
This will only occur if a raycast done on the initial mouse click is valid, as without any raycast result there is no location selected or targeted.