FEATURE - Create Unit Spawner component
This commit creates the Unit Spawner component, which can be used to spawn units both periodically and on command.
It has several parameters which can be used to configure it, including the types of units it can periodically spawn and their corresponding intervals, the fixed offset from the spawner unit's position where units will be spawned, the boundary surrounding the spawner where units can be spawned, and the number of times spawning can be retried when the target location is occupied before failing.
A crude random algorithm has been defined to pick a new spawn position when the current spawn position is occupied. This should be refined further later on, or replaced entirely in favor of units spawning on top of each other resolving this smoothly (both sliding away from each other or whatever)

There is a parameter for toggling 'random mode' but this currently does nothing.

A Test Spawner prefab has been added to demonstrate.