FEATURE - Add Targeting Component
The Targeting Component is responsible for tracking enemy units and determining what should be the current target. It also allows for a target to be set via player command via the _orderedTarget parameter, which will override other targeting
Due to performance concerns, had to tweak algorithm to only automatically update the current target with a smaller radius targeting check, with the full radius targeting check only being done on request. This will probably lead to problems down the line, but the ranges can be tweaked as needed.
Just keep in mind a large radius will cause major slowdown if there are hundreds (or perhaps only like 50-100) of other units within the targeting range.
A potential future alternative to this would be to raycast in all directions surrounding the unit, but this has its own performance implications.