FEATURE - Attack Component part 1 - basic skeleton
The Attack Component will handle controlling the unit's attacking based on the current state, available targets, and weapon in use.
This commit starts on some of the methods, includes most of the other components necessary to use the attacking component, and provides an outline on all methods necessary for the attacking component.
Some additional features might be needed in other components, what I can think of right now
-Movement component: Need to have movement component follow target somehow. Maybe it can be linked to the targeting component, and then all the attacking component will need to do is call a method to make it follow the target?
-Unit State: Might need more complex states to cover all scenarios with the attack component, but this is tbd, we shall see if this is fine as is or not