FEATURE - Health Component part 1 - everything except for projectile/death handling
The Health Component, when completed, will monitor and manage the unit's health and defense, handle taking damage from enemy projectiles, and detect + report when the unit is dead so it can be destroyed.
Currently, the projectile handling is unfinished as the projectiles do not exist, and the death reporting is unfinished but can be completed soon.