REFACTOR - Set up supported component list of unit via Unit Database
Previously, the Unit Info Component and Unit Database both initialized separate lists of components.
This is dumb and makes zero sense, so to make my approach to this less cumbersome, the Unit Creation Controller will now set the supported components list in the Unit Info component using the Unit Database.
This may or may not interfere with work in the other scenes.