FEATURE - Create Worker unit, set up harvesting ability - part 1
The Worker unit will be responsible for harvesting resources and constructing buildings. For now, this commit enables it to do the former with the following changes.
Harvesting component created - enables worker to harvest a target resource which is sent by the controller, after arriving at the deposit
Movement component modified to support reporting when it arrives at the destination, and support an order to move to harvest a certain resource with a special state
Capability Model and Order Model have respective additions to support the harvesting action and harvest order
-Unit Controller updated to support handling of harvest order
-Unit Creation Controller updated to set up callback from Harvesting component to event handler for ResourceHarvestEvent (report newly harvested resource to controller)

Still to do: Update Game State Controller and add Game State Model so that player resources are tracked and updated.