FIX - fix resource harvesting continuing after unit moved away from deposit
Previously, was not possible for movement component to transition the unit state into "moving" unless it was previously in "idle".
Since resource harvesting takes place in the "harvesting" state, this meant the unit would continue to harvest even after being moved off of the resource deposit, as long as the deposit was not depleted.
This commit stops this by allowing a transition into "moving" state from "harvesting" state.