FEATURE - Create Civilian Building
The Civilian Building will hold civilians, and spawn workers. The worker spawning is limited by a cap that is calculated based on the number of civilians.
The player will be able to issue a command to this building to start evacuating its civilians, which will cause the civilian population in the building to diminish over time.
Adds the evacuateCivies capability, and EVAC_CIVIES order to support this functionality.