FEATURE - skeleton for MenuSelectionEvent stuff
Includes definition and callback function for MenuSelectionEvent
Includes DisplayInfoController features which provides functions for getting additional information to make available to the UI, clearing that information, and an interface for the UI to access that information. Comments in the code add further detail. This commit might be missing something because I'm tired....