FEATURE - Worker can construct buildings - rough draft
The worker unit must be able to construct buildings. This commit adds a Construction component to support this functionality on the worker, and also adds a Unit Spawner to the worker to facilitate creation of the building.

The Capability Model is updated to include the 'construct' capability.
The Order Model is updated to include  the 'ORDER_SELECT_BUILDING' and 'ORDER_CONSTRUCT' orders. Note that it is advisable in the future to add Capability Model support for multiple capabilities bound to a single action, to avoid these orders supporting the 'same' action (different behavior for targeted/untargeted versions)

Unit Controller updated to support selecting the building to construct, and moving a worker to construct a building.
Display Info Controller updated to add additional menu options representing the buildings the worker can construct when the 'construct' ability button is pressed.
Movement component given new method to order movement to construct a building.
Added unit state for moving to construct.