FEATURE - Attack component part 1.5
Organized the weapon handling, and method calls for state transitions. Also added a boolean that tracks whether the target is in range or not to support "in range"/"out of range" state changes.
The hard parts still remain: movement component features to support rotating in place without breaking existing functionality, and move towards a specific target rather than position (probably not hard, but the state transitions will be annoying)