FIX - deal with "construct" error, rename DisplayAdditionalInfo method
real reason the menu buttons were not appearing was because MenuSelectionEvent is not sent when "construct" is selected. This will need to be a UI-side fix to call the correct event handler method.
I did confirm that the construct action shows up when worker selected (3rd button), but again, clicking it triggers UIOrderEvent