FIX - Update Mouse Controller to avoid handling mouse clicks in "UI Regions"
The Mouse Controller will now check if a mouse is in any of the predefined "UI regions" - regions on the screen occupied by the UI, and avoid sending a SelectionEvent or MouseOrderEvent if it is.
The current hard-coded storing of these UI-regions is poor design and will have to be updated. Either a better method of detecting UI elements, or a refactor to dynamically compute this in the UI controller, then send it to the mouse controller will have to be done, in order to avoid tedious updates to the UI regions when dimensions change, and allow for scaling the UI.