FIX - Fix Attack Component not resuming action after being ordered to look away from target
If ordered movement causes a unit to look away from its target, without moving out of range of the unit, and without moving closer to a new target, the unit will not be able to resume action after the ordered movement has finished.
Therefore, need to check if ordered movement is in progress if the unit is unable to fire despite being in range, with the same handling as before following when the ordered movement ends.