FEATURE - Attack component handling for collision with target
Due to the sizes of some units, determining when a unit is in range for a melee attack can prove unreliable, so instead, we automatically consider a unit in range if it collides with its target.
One catch, due to the reliance on OnCollisionEnter, if a unit is colliding with 2+ potential targets, it will not properly update the _collidedWithTarget flag when switching targets. Using OverlapSphere with a small radius instead is likely far more effective, but this will stand for now.