FEATURE - add "selection rings" that highlight selected units
When a unit is selected, the Selection Controller will assign a Selection Indicator to it that highlights its existence.
Might be best to use the scifi particle system asset instead of the shader I came up with due to the quad used by it interfering with illumination, even with alpha at 0.
Might also make more sense to manage these in the UI controller long-term.