FEATURE - Add Area Selection Functionality
This commit allows for the selection of multiple units by clicking and dragging over a specific region.
Due to the orientation of the camera, the original idea of using an AABB test to determine if units were in the box did not work, and a very scuffed OBB test has to be done instead.
Note: This can cause crashes due to the UI only having 6 ability slots, if the wrong combination of units are selected (ex: worker and infantry unit at same time)