FEATURE - provide interface for sending the current mouse positions for an area selection to the UI
The UI will use these mouse positions to draw a box on the screen. Not sure If I actually want this to be the approach since these mouse positions will not represent the selection accurately if the player moves the camera. Probably best to add some world-space effect long term... will think about this more tomorrow