FEATURE - Hermite Spline Movement initial update
NOTE: dynamic destination is currently disabled, so until it is re-enabled, the attack component will not work as intended.
Implements hermite spline movement, but with the catch of still needing rigidbody movement methods for the physics to work properly.
Speed tuning, ease in/ease out, and dynamic destination compatibility still needed.