FEATURE - update hermite spline movement to take into account unit speed and path distance
Unit's rate of progression through the spline now based on the speed parameter of the unit, and the (approx) length of the path. Unit movement should now be largely accurate.
Some issues remain due to NavMesh being inaccurate, probably need to make proper use of NavMeshAgent and NavMeshObstacle to fix.