FIX - Semi-temporary fix for problem with using pathfinding to move towards obstacles
Problem: If a movement target is a unit with a NavMeshObstacle component which is carving the mesh, then the pathfinding will fail (probably because the destination is not on the navmesh? idk).
The temporary solution is the FindUnobstructedPath method, which will try an alternative destination which is not directly centered on the obstacle in an effort to get the path planning to work properly.
Do not like this solution but it works for now...
In addition: Re-enabled dynamic destination so the Attack component is now working again, but issues remain.