FEATURE - Add Unit Health Bar
Note: Currently doesn't show up for civilian buildings because the y-axis scaling is screwed up for other models (in terms of root localScale), so my adjustment to the health bar positioning causes it to fall underneath the properly scaled civilian building...

Added a health bar that will the health of the attached unit. The health bar is linked to an associated unit upon creation and will stick with it until it dies. Better approach might be to insert the health bar into the hierarchy of the unit, allowing for better y-axis control? Or just hardcode a special case for civilian building lmao...