FEATURE - Initial update for delayed death
Updated UnitCreationController to only kill the unit after 2 seconds.
Requires some very scuffed handling for now, will 1000% want to refactor, this includes literally deleting the UnitInfo component so the unit doesn't register to other components as a valid unit, which required extra updating to the projectile class to prevent a crash....
will update this to be done properly later, but this will allow for death animation to play for now.