FEATURE - Add parsing of rule file in AIControl
AI Control will now parse its given rule file following the design proposed in the onenote. Keep in mind the order in which or/and statements are split for prereqs.
Also added debug mode, and debug flag to rules for easier troubleshooting. Debug mode is enabled in the inspector for the component, and local debug mode is enabled for a rule by adding the 'D' flag to it.