REFACTOR - Remove all "ordered movement" features in the Movement component
'Ordered movement' distinction is no longer necessary with the advent of the AI Controller.
This eyesore which has caused nothing but problems can finally be nuked, and replaced because the AI Controller will use "commands" to grant player-issued movement commands priority over automatic behaviour.
Had to do some misc stuff in the Attack component and Unit Controller to get it to compiler, none of that stuff matters because everythings getting refactored to kingdom come in those components to work with the AI Control