FEATURE - Add dynamic calculation of UI-occupied regions by UI Controller
UI Controller will now include a list in the inspector of elements that will obstruct the game when active. Upon initialization, the boundaries will be calculated in the same format (percentages of screen) that was used for the hardcoded version in MouseController.
MouseController will now get these bounding regions to determine if the UI is obstructing the mouse click. If one or more of the obstructing UI elements is inactive, it will not be included.