FEATURE - Enemy AI part 1 - Commander and CommandableUnit scripts
For the enemy AI, edenite ravagers will be taking control of edenite munchers to do their bidding. This requires a new Commander script for acquiring and controlling commanded units, and a CommandableUnit script to manage control over a specific unit.
Added a query in the Targeting component for getting units of a certain type within a specified range