FEATURE - Enemy AI part 4 - Add edenite ravager retreating behavior
When the Edenite Ravager has lost too many of the units under its command, it will attempt to retreat back to its spawn region to gather more units.
Updated the Movement component to add a getter for the return point so the ravager can determine if it is in its spawn region.
Updated the Targeting component to provide functionality for resetting the targetting state, as due to attacking rules being largely ignored during retreat, the targeting state will be stale once the unit has finished retreating if a player unit has remained in targeting range.