FIX - fix health component reporting low health mutliple times
This also "fixed" the default movement/pathfinding transition issue in the Edenite Devil. Turns out this issue is caused when a unit that is already moving towards a target, and switches into default movement mode, carries out the moveTarget action again. This will cause it to try and pathfind again, but because it is so close to the NavMeshObstacle target, it gets no real path, and it never switches into default mode because it has to carry out moveTarget each frame, so the unit is frozen in place.